Move searchsorted* functions to SortedSearch stdlib module
Keep them under Base.Sorted but unexported, since they are used
in a few places in Base, in particular for sparse vectors and matrices.
