+++ base/linalg/svd.jl
@@ -253,7 +253,7 @@ svdvals(S::SVD{<:Any,T}) where {T} = (S[:S])::Vector{T}
 
 # SVD least squares
 function ldiv!(A::SVD{T}, B::StridedVecOrMat) where T
-    k = searchsortedlast(A.S, eps(real(T))*A.S[1], rev=true)
+    k = Base.Sort.searchsortedlast(A.S, eps(real(T))*A.S[1], rev=true)
     view(A.Vt,1:k,:)' * (view(A.S,1:k) .\ (view(A.U,:,1:k)' * B))
 end
