+++ doc/src/index.md
@@ -76,6 +76,7 @@
   * [File Events](@ref lib-filewatching)
   * [Iterative Eigensolvers](@ref lib-itereigen)
   * [Unicode](@ref)
+  * [SortedSearch](@ref)
 
 ## Developer Documentation
