+++ doc/src/manual/arrays.md
@@ -273,10 +273,13 @@ julia> x[1, [2 3; 4 1]]
 ```
 
 Empty ranges of the form `n:n-1` are sometimes used to indicate the inter-index location between
-`n-1` and `n`. For example, the [`searchsorted`](@ref) function uses this convention to indicate
+`n-1` and `n`. For example, the [`searchsorted`](@ref) function from the standard library
+module `SortedSearch` uses this convention to indicate
 the insertion point of a value not found in a sorted array:
 
 ```jldoctest
+julia> using SortedSearch
+
 julia> a = [1,2,5,6,7];
 
 julia> searchsorted(a, 3)
