+++ doc/src/stdlib/.gitignore
@@ -8,3 +8,4 @@ filewatching.md
 crc32c.md
 dates.md
 unicode.md
+sortedsearch.md
