+++ stdlib/SortedSearch/docs/src/index.md
@@ -1,0 +1,7 @@
+# SortedSearch
+
+```@docs
+SortedSearch.searchsorted
+SortedSearch.searchsortedfirst
+SortedSearch.searchsortedlast
+```
