+++ stdlib/SortedSearch/src/SortedSearch.jl
@@ -1,0 +1,11 @@
+# This file is a part of Julia. License is MIT: https://julialang.org/license
+
+__precompile__(true)
+
+module SortedSearch
+
+using Base.Sort: searchsorted, searchsortedfirst, searchsortedlast
+
+export searchsorted, searchsortedfirst, searchsortedlast
+
+end
